Add emacs-style tiling: binary split tree, multi-tile rendering, per-tile scissor clipping
- Binary split tree (TileNode/TileLeaf/TileSplit) with per-tile widget state
- TUI: cell-bordered tiles with per-tile status bars, inactive frame caching
- Metal: pixel-line borders, per-tile scissor rects, full scroll support
- Lisp API: split-window-right/below, delete-window, other-window, etc.
- C-x 2/3/0/1/o keybindings
- Horizontal + vertical scroll shifts text and widgets together
- Hit detection accounts for both text and widget scroll offsets
- HitGrid uses full descendant extent for overflowing layouts
- Inactive tile frame caching (skip recomputation when unchanged)
- Timeline lane labels rendered as Metal GlyphRun primitives

+++ src/widget_render/label.rs
@@ -94,8 +94,8 @@ impl WidgetDefinition for LabelWidget {
                 color: bg,
             }),
             MetalPrimitive::GlyphRun(MetalGlyphRunPrimitive {
-                row: node.rect.row,
-                col: node.rect.col,
+                row: node.rect.row as i32,
+                col: node.rect.col as i32,
                 text: text.clone(),
                 fg,
                 bg,
